server/requirements.txt
fastapi==0.120.1
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.28.1
jose==1.0.0
passlib==1.7.4
pdfplumber==0.11.0
pydantic==2.12.3
pymongo==4.15.3
python-dotenv==1.2.1
python_docx==1.2.0
python_jose==3.5.0